automations/keyword_ranking_tracker/requirements.txt
requests
pyyaml
pandas
gspread # optional for Google Sheets integration
oauth2client # required for gspread
